pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tomato-api"
version = "1.0.0"
description = "Python API to communicate with Tomato Router"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Steven Aubertin"}
]
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Networking",
]
keywords = ["tomato", "router", "network", "dhcp", "arp"]
dependencies = [
    "requests>=2.25.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-mock>=3.10.0",
    "pytest-cov>=4.0.0",
    "responses>=0.23.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "types-requests>=2.31.0",
]

[project.scripts]
tomato-devlist = "src.devlist:main"
tomato-staticlist = "src.staticlist:main"
tomato-unknown = "src.unknown_devices:main"
tomato-find = "src.find_device:main"
tomato-hostname = "src.hostname_lookup:main"

[project.urls]
Homepage = "https://github.com/stevenaubertin/tomato-api"
Repository = "https://github.com/stevenaubertin/tomato-api"
Issues = "https://github.com/stevenaubertin/tomato-api/issues"

[tool.setuptools]
packages = ["src", "src.routers"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]
filterwarnings = [
    "ignore::urllib3.exceptions.InsecureRequestWarning",
]

[tool.coverage.run]
source = ["src"]
branch = true
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]
show_missing = true

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
files = ["src"]

[[tool.mypy.overrides]]
module = ["responses.*", "dotenv.*", "urllib3.*", "requests.*"]
ignore_missing_imports = true

[tool.ruff]
target-version = "py38"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
